+++ source/B2B.Transactions.CimMessageAdapter/Messages/MarketActivityRecord.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace B2B.CimMessageAdapter
+namespace B2B.CimMessageAdapter.Messages
 {
     public class MarketActivityRecord
     {
